+++ docs/source/docs/w90_files.rst
@@ -1,12 +1,12 @@
 Classes to store Wannier90 input and output files
 ---------------------------------------------------
 
-.. autoclass:: wannierberri.system.Wannier90data
+.. autoclass:: wannierberri.w90files.Wannier90data
    :members: __init__, disentangle, get_disentangled, write
    :show-inheritance:
 
 
-.. automodule:: wannierberri.system.w90_files
+.. automodule:: wannierberri.w90files
    :undoc-members:
    :members: CheckPoint, CheckPoint_bare, WIN, AMN, MMN, UXU, UHU, UIU, SXU, SHU, SIU, SPN
    :show-inheritance:
