+++ wannierberri/wannierise/__init__.py
@@ -1,0 +1,1 @@
+from .disentanglement import disentangle
